refact(main): split main in __main__.py file
