+++ setup.py
@@ -53,8 +53,8 @@ setup(
     author_email='rich@openwatch.net',
     entry_points={
         'console_scripts': [
-            'omnihash = omnihash.__init__:main',
-            'oh = omnihash.__init__:main',
+            'omnihash = omnihash.__main__:main',
+            'oh = omnihash.__main__:main',
         ],
         'omnihash.plugins': [
             'a_sha3 = omnihash.plugin:plugin_sha3_digesters [sha3]',
